requirements.txt
streamlit
biopython
numpy
scipy
toytree
matplotlib
plotly